Fix BFloat16 numpy conversion error in Hungarian matcher
The scipy.optimize.linear_sum_assignment requires numpy arrays, but
BFloat16 dtype is not supported by numpy. This fix converts the cost
matrix to float32 before converting to numpy array.

Error fixed:
TypeError: Got unsupported ScalarType BFloat16

+++ hungarian_loss.py
@@ -72,7 +72,8 @@ class HungarianMatcher(nn.Module):
 
             # Total cost
             C = self.cost_class * cost_class
-            C = C.cpu().numpy()
+            # Convert to float32 first (BFloat16 is not supported by numpy)
+            C = C.float().cpu().numpy()
 
             # Solve optimal assignment using Hungarian algorithm
             src_idx, tgt_idx = linear_sum_assignment(C)
